created mixin for django-q tests
